Added a look command that allows the player to "pause" the game, and (#9)
look around the map via a cursor that uses the same controls as moving
the player. When in look mode, the player can examine entities, and gets
info on map tiles as well. This also necessitated a new description
system for non-item entities, basically a description component that
gives a brief description of an entity, and then an auto-generated
description based on its stats, if it has any, and its health/mana/etc.
This produces a nice way for the player to view the world around them,
so they can react accordingly.