Effects system (#10)
* Added basic effects system, and the ability to apply effects to an
entity. Also, consume consumables, which applies an effect to an entity.

* Added targeting system, this will be used for throwing and spell
casting.

* Added throwing, and ground pools for items that contain liquids
(potions, poisons, etc). Consumables can be thrown to a location
(determined by strength), and shatter on impact. If they contain
liquids, they create a ground pool, that confers the effect of the
liquid on anything that is standing in the pool.

* Added ability to target other entities with consumables (for example,
throwing a potion at an entity will apply the potions effects to the
target entity, should allow for more dynamic gameplay)

* Updated consumables to have timers, as well as consumables that can
affect an entities stats. Also added a cleaner way to define
consumables. This completes the consumable/effect/throwing system.

* Updated readme